Move ripped THREE js utilities into their own file
